Added DayView and WeekView Tabs to GUI (and other GUI tweaks)
Changes:
--> Reworked the dayView tab to allow the user to scroll across the
hours of the day. (Will eventually add events to these views dependent
on the selected date).
--> Added weekView tab to allow the user to scroll across the hours of
the day, for an entire week. (Will eventually add events to these views
dependent on the selected date).
--> Fixed error with logging out/switching user and not being able to
log back in again.
--> Ensured that the application window is centred on the screen each
time the scene is switched from LoginView to CalendarView.
--> Reduced the width of the "Upcoming Events" section to allow for more
room for the weekView tab (and the month tab, as well).