Added dumpEvents to User class
'dumpEvents' is an instance method of the User class which searches
though all the events saved in a user save file and then prints them to
the console.

Triggered by the "Admin Preferences" menu button in CalendarView.

This is useful for debugging purposes, as it allows you to see the
events present in the user save file.