Added updateEventPane and addEventBlock Functions
Changes:
--> 'updateEventPane' takes a ZonedDateTime date and an AnchorPane as
input to find events for that specific date, then creates and adds
visual event blocks for that specific event to the specified pane (via
addEventBlock).
--> 'addEventBlock' takes an event and an Anchor pane as input to create
visual event blocks (Rectangle) and add them to the specified pane.

Issues:
--> Unsure whether this is an issue inside of CalendarView or as part of
the TimeUnit subclasses, but the day of week does not seem to correlate
properly when creating a new event in EventMakerView.
--> Adding event data on top of the eventBlock Rectangle?
--> *** Need to implement different process for InstantEvent. Only
TimedEvents are compatible.